Apply review names.
