+++ src/Microsoft.DotNet.XHarness.TestsRunners.Common/ApplicationOptions.cs
@@ -8,7 +8,7 @@ using Microsoft.DotNet.XHarness.Common;
 using Microsoft.DotNet.XHarness.iOS.Shared.Execution.Mlaunch;
 using Mono.Options;
 
-namespace Microsoft.DotNet.XHarness.Tests.Runners.Core
+namespace Microsoft.DotNet.XHarness.TestsRunners.Common
 {
 
     internal enum XmlMode
