+++ src/Microsoft.DotNet.XHarness.TestsRunners.Common/MinimumLogLevel.cs
@@ -2,13 +2,15 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace Microsoft.DotNet.XHarness.Tests.Runners.Core
+namespace Microsoft.DotNet.XHarness.TestsRunners.Common
 {
-    internal static partial class Extensions
+    public enum MinimumLogLevel
     {
-        public static string YesNo(this bool b)
-        {
-            return b ? "yes" : "no";
-        }
+        Critical,
+        Error,
+        Warning,
+        Info,
+        Debug,
+        Verbose
     }
 }
