+++ src/Microsoft.DotNet.XHarness.TestsRunners.Common/TestAssemblyInfo.cs
@@ -5,7 +5,7 @@
 using System;
 using System.Reflection;
 
-namespace Microsoft.DotNet.XHarness.Tests.Runners.Core
+namespace Microsoft.DotNet.XHarness.TestsRunners.Common
 {
     public class TestAssemblyInfo
     {
