+++ src/Microsoft.DotNet.XHarness.TestsRunners.Common/TestCompletionStatus.cs
@@ -2,15 +2,14 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace Microsoft.DotNet.XHarness.Tests.Runners.Core
+namespace Microsoft.DotNet.XHarness.TestsRunners.Common
 {
-    public enum MinimumLogLevel
+    internal enum TestCompletionStatus
     {
-        Critical,
-        Error,
-        Warning,
-        Info,
-        Debug,
-        Verbose
+        Undefined,
+        Passed,
+        Failed,
+        Skipped,
+        Inconclusive,
     }
 }
