+++ src/Microsoft.DotNet.XHarness.TestsRunners.Common/TestExecutionState.cs
@@ -4,7 +4,7 @@
 
 using System;
 
-namespace Microsoft.DotNet.XHarness.Tests.Runners.Core
+namespace Microsoft.DotNet.XHarness.TestsRunners.Common
 {
     internal class TestExecutionState
     {
