+++ src/Microsoft.DotNet.XHarness.TestsRunners.Common/TestResult.cs
@@ -1,4 +1,4 @@
-﻿namespace Microsoft.DotNet.XHarness.Tests.Runners.Core
+﻿namespace Microsoft.DotNet.XHarness.TestsRunners.Common
 {
     /// <summary>
     /// Enumeration used to state the result of a test.
