+++ src/Microsoft.DotNet.XHarness.TestsRunners.Xunit/Properties/AssemblyInfo.cs
