+++ src/Microsoft.DotNet.XHarness.TestsRunners.Xunit/XUnitFilterType.cs
@@ -2,7 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace Microsoft.DotNet.XHarness.Tests.Runners.Xunit
+namespace Microsoft.DotNet.XHarness.TestsRunners.Xunit
 {
     internal enum XUnitFilterType
     {
