+++ src/Microsoft.DotNet.XHarness.TestsRunners.Xunit/iOSApplicationEntryPoint.cs
@@ -4,9 +4,9 @@
 
 using System;
 using System.Threading.Tasks;
-using Microsoft.DotNet.XHarness.Tests.Runners.Core;
+using Microsoft.DotNet.XHarness.TestsRunners.Common;
 
-namespace Microsoft.DotNet.XHarness.Tests.Runners.Xunit
+namespace Microsoft.DotNet.XHarness.TestsRunners.Xunit
 {
     public abstract class iOSApplicationEntryPoint : ApplicationEntryPoint
     {
