+++ tests/Microsoft.DotNet.XHarness.Tests.Runners.Tests/Microsoft.DotNet.XHarness.Tests.Runners.Tests.csproj
@@ -21,8 +21,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\Microsoft.DotNet.XHarness.Tests.Runners.Core\Microsoft.DotNet.XHarness.Tests.Runners.Core.csproj" />
-    <ProjectReference Include="..\..\src\Microsoft.DotNet.XHarness.Tests.Runners.Xunit\Microsoft.DotNet.XHarness.Tests.Runners.Xunit.csproj" />
+    <ProjectReference Include="..\..\src\Microsoft.DotNet.XHarness.TestsRunners.Common\Microsoft.DotNet.XHarness.TestsRunners.Common.csproj" />
+    <ProjectReference Include="..\..\src\Microsoft.DotNet.XHarness.TestsRunners.Xunit\Microsoft.DotNet.XHarness.TestsRunners.Xunit.csproj" />
   </ItemGroup>
 
 </Project>
