+++ tests/Microsoft.DotNet.XHarness.Tests.Runners.Tests/xUnit/XUnitFilterTests.cs
@@ -7,8 +7,8 @@ using System;
 using System.Collections.Generic;
 using System.Reflection;
 using System.Text;
-using Microsoft.DotNet.XHarness.Tests.Runners.Core;
-using Microsoft.DotNet.XHarness.Tests.Runners.Xunit;
+using Microsoft.DotNet.XHarness.TestsRunners.Common;
+using Microsoft.DotNet.XHarness.TestsRunners.Xunit;
 using Moq;
 using Xunit;
 using Xunit.Abstractions;
